feat: 삭제하기 기능 완성(나중에 합치고 함수 동작 변경 필요)
