+++ src/api/answerApi.js
@@ -83,20 +83,24 @@ export const putUpdateAnswer = async (answerId, { content, isRejected }) => {
 //   // answerId, questionId, 수정내용, isRejected, createdAt줌
 // }
 
-// // 답변 삭제하는 기능
-// export async function DeleteAnswer(answerId) {
+// // 답변 삭제하는 기능 (질문삭제하면 답변도 같이 삭제됨)
+// export const DeleteAnswer = async (answerId) => {
 //   const res = await axios.delete(
 //     `https://openmind-api.vercel.app/3-7/answers/${answerId}/`,
 //   );
-//   const result = res.json();
+//   const data = res;
+//   console.log(data);
+//   return data;
 //   // result로 삭제 성공여부 받기
-// }
+// };
 
-// // 질문 삭제하는 기능
-// export async function DeleteQuestion(answerId) {
-//   const res = await axios.delete(
-//     `https://openmind-api.vercel.app/3-7/questions/${questionId}/`,
-//   );
-//   const result = res.json();
-//   // result로 삭제 성공여부 받기
-// }
+// 질문 삭제하는 기능
+export const DeleteQuestion = async (questionId) => {
+  const res = await axios.delete(
+    `https://openmind-api.vercel.app/3-7/questions/${questionId}/`,
+  );
+  const data = res;
+  console.log(data);
+  return data;
+  // result로 삭제 성공여부 받기
+};
